refactor: colocate integration tests with the source modules they cover
Moves integration test files from __tests__/integration/ into __tests__/app/
mirroring the source tree, so each module's integration test lives next to
the module it exercises:

  app/modules/cloud-code/crud/            -> crud/index.integration.test.js
  app/modules/cloud-code/crud/custom-queries/ -> custom-queries/index.integration.test.js
  app/modules/cloud-code/aggregate/       -> aggregate/index.integration.test.js
  app/epics/FormCreator/                  -> FormCreator/formCreator.integration.test.js
  app/modules/cloud-code/user/            -> user/auth.integration.test.js

Infrastructure (globalSetup, globalTeardown, parseClient, cloudCode) stays in
__tests__/integration/. jest.integration.config.js: testMatch broadened to
__tests__/**/*.integration.test.js, maxWorkers:1 added to prevent data races
on the shared Parse Server. jest.config.js: excludes *.integration.test.js
from the unit test run.
