+++ __tests__/app/epics/FormCreator/formCreator.integration.test.js
@@ -9,7 +9,7 @@
  *   yarn test:integration
  */
 
-const { initParse, Parse } = require('../parseClient');
+const { initParse, Parse } = require('../../../integration/parseClient');
 
 beforeAll(() => {
   initParse();
